dsrp_ml_utils/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dsrp-ml-utils"
version = "0.1.0"
description = "ML utilities for DSRP Machine Learning Engineering course - movie recommendation pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "DSRP", email = "info@dsrp.dev" }
]
keywords = ["machine-learning", "mlops", "recommendation-system", "dsrp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "polars>=0.20.0",
    "mlflow>=2.0.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
azure = [
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
all = [
    "dsrp-ml-utils[azure,dev]",
]

[project.urls]
Homepage = "https://github.com/dsrp-org/dsrp-machine-learning-engineering-4"
Repository = "https://github.com/dsrp-org/dsrp-machine-learning-engineering-4"
Documentation = "https://github.com/dsrp-org/dsrp-machine-learning-engineering-4#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/dsrp_ml_utils"]

[tool.hatch.build.targets.sdist]
include = [
    "src/dsrp_ml_utils/**/*.py",
    "README.md",
    "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
